move model pretty printer into its own module
